test: 그룹 인가 테스트 하나로 합치기 (#352)
