+++ src/test/java/com/exchangediary/global/config/interceptor/GroupAuthorizationInterceptorTest.java
@@ -4,9 +4,9 @@ import com.exchangediary.ApiBaseTest;
 import com.exchangediary.group.domain.GroupRepository;
 import com.exchangediary.group.domain.entity.Group;
 import com.exchangediary.member.domain.MemberRepository;
-import com.exchangediary.member.domain.entity.Member;
 import com.exchangediary.member.domain.enums.GroupRole;
 import io.restassured.RestAssured;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -20,9 +20,73 @@ public class GroupAuthorizationInterceptorTest extends ApiBaseTest {
     private MemberRepository memberRepository;
 
     @Test
+    void 그룹_가입하지_않은_사용자가_그룹생성가입페이지_접근() {
+        RestAssured
+                .given().log().all()
+                .cookie("token", token)
+                .redirects().follow(false)
+                .when().get("/group")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value());
+    }
+
+    @Test
+    void 그룹_가입한_사용자가_그룹월별형페이지_접근() {
+        Group group = createGroup();
+        updateSelf(group);
+
+        RestAssured
+                .given().log().all()
+                .cookie("token", token)
+                .redirects().follow(false)
+                .when().get("/group/" + group.getId())
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value());
+    }
+
+    @Test
+    void 그룹_가입한_사용자가_그룹생성가입페이지_접근() {
+        Group group = createGroup();
+        updateSelf(group);
+
+        String location = RestAssured
+                .given().log().all()
+                .cookie("token", token)
+                .redirects().follow(false)
+                .when().get("/group")
+                .then()
+                .log().status()
+                .log().headers()
+                .statusCode(HttpStatus.FOUND.value())
+                .extract()
+                .header("location");
+
+        Assertions.assertThat(location.substring(location.indexOf("/group"))).isEqualTo("/group/" + group.getId());
+    }
+
+    @Test
+    void 그룹_가입하지_않은_사용자가_그룹월별형페이지_접근() {
+        Group group = createGroup();
+
+        String location = RestAssured
+                .given().log().all()
+                .cookie("token", token)
+                .redirects().follow(false)
+                .when().get("/group/" + group.getId())
+                .then()
+                .log().status()
+                .log().headers()
+                .statusCode(HttpStatus.FOUND.value())
+                .extract()
+                .header("location");
+
+        Assertions.assertThat(location.substring(location.indexOf("/group"))).isEqualTo("/group");
+    }
+
+    @Test
     void 그룹_속한_멤버_api_접근 () {
         Group group = createGroup();
-        updateSelfInfo(member, group);
+        updateSelf(group);
 
         RestAssured
                 .given().log().all()
@@ -37,7 +101,7 @@ public class GroupAuthorizationInterceptorTest extends ApiBaseTest {
     void 그룹_속하지_않은_멤버_api_접근 () {
         Group group = createGroup();
         Group otherGroup = createGroup();
-        updateSelfInfo(member, group);
+        updateSelf(group);
 
         RestAssured
                 .given().log().all()
@@ -50,7 +114,7 @@ public class GroupAuthorizationInterceptorTest extends ApiBaseTest {
     @Test
     void 그룹_속한_멤버_일기에_접근 () {
         Group group = createGroup();
-        updateSelfInfo(member, group);
+        updateSelf(group);
 
         RestAssured
                 .given().log().all()
@@ -65,7 +129,7 @@ public class GroupAuthorizationInterceptorTest extends ApiBaseTest {
     void 그룹_속하지_않은_멤버_일기에_접근() {
         Group group = createGroup();
         Group otherGroup = createGroup();
-        updateSelfInfo(member, group);
+        updateSelf(group);
 
         String location = RestAssured
                 .given().log().all()
@@ -85,7 +149,7 @@ public class GroupAuthorizationInterceptorTest extends ApiBaseTest {
     @Test
     void 그룹_속한_멤버_월별형_접근 () {
         Group group = createGroup();
-        updateSelfInfo(member, group);
+        updateSelf(group);
 
         RestAssured
                 .given().log().all()
@@ -100,7 +164,7 @@ public class GroupAuthorizationInterceptorTest extends ApiBaseTest {
     void 그룹_속하지_않은_멤버_월별형_접근() {
         Group group = createGroup();
         Group otherGroup = createGroup();
-        updateSelfInfo(member, group);
+        updateSelf(group);
 
         String location = RestAssured
                 .given().log().all()
@@ -121,14 +185,14 @@ public class GroupAuthorizationInterceptorTest extends ApiBaseTest {
         return groupRepository.save(Group.of("group-name", "code"));
     }
 
-    private Member updateSelfInfo(Member member, Group group) {
-        member.joinGroup(
-                "api요청멤버",
-                "orange",
+    private void updateSelf(Group group) {
+        this.member.joinGroup(
+                "me",
+                "red",
                 1,
                 GroupRole.GROUP_MEMBER,
-                group);
-        memberRepository.save(member);
-        return member;
+                group
+        );
+        memberRepository.save(this.member);
     }
 }
